ci(python-autogame:御魂协战): 御魂协战
小号无法单挑魂十，考虑依旧以魂十进行协战

+++ python/autogame/src/service/image_service.py
@@ -95,7 +95,7 @@ class ImageService:
         return False
 
     @staticmethod
-    def wait(folder_path: str, cvstrategy: [] = CVSTRATEGY, timeout: int = 20,
+    def wait(folder_path: str, cvstrategy: [] = CVSTRATEGY, timeout: float = 20,
              threshold: float = THRESHOLD, interval: float = INTERVAL, is_throw: bool = False,
              is_click: bool = False):
         """
